main 마크다운 로직 변경
